requirements.txt
transformers
groq
easyocr
python-Levenshtein
streamlit
numpy